use script inside of plugin instead of thread executor
